refactor grids module
